Spectral-Extension/requirements.txt
torchvision
torchsummary
matplotlib
rasterio
geopandas
pystac_client
planetary_computer
odc-stac
scikit-image
lmdb
opencv-python
tensorboard
tqdm
tensorflow
tensorboardX